Add SeasonSummary module and include in Season class
